rename excel_styles to excel_styles_and_validation
